+++ core/indexer/src/reactor/lite_executor.rs
@@ -138,7 +138,7 @@ impl LiteExecutor {
                 )
                 .await;
             let payment = indexer_types::Payment {
-                signer_id: signer.signer_id().expect("test signer must have id") as u64,
+                signer_id: signer.signer_id().expect("test signer must have id"),
                 gas_limit: runtime.gas_limit_for_non_procs,
             };
             runtime
@@ -176,8 +176,8 @@ impl Executor for LiteExecutor {
     async fn execute_transaction(
         &self,
         runtime: &mut Runtime,
-        height: i64,
-        tx_id: i64,
+        height: u64,
+        tx_id: u64,
         tx: &indexer_types::Transaction,
     ) -> anyhow::Result<Vec<Vec<Option<anyhow::Error>>>> {
         runtime
@@ -196,7 +196,7 @@ impl Executor for LiteExecutor {
             .await;
 
         let payment = indexer_types::Payment {
-            signer_id: self.signer.signer_id().unwrap_or(0) as u64,
+            signer_id: self.signer.signer_id().unwrap_or(0),
             gas_limit: runtime.gas_limit_for_non_procs,
         };
         match runtime
